Add support git{hub|lab} urls, username settings
Add support for shorter github and gitlab url, by skipping the schema
and @ part.

Add support for github.username and gitlab.username settings. If the
setting matches the account name in git*.com/<account>/<repository> then
the url is turn into git@git*.com:<account>/<repository>, otherwise, it is
completed as https://git*.com/<account>/<repository>.

Add a description of the commands in the readme.